blog_bot/requirements.txt
anthropic>=0.26.0
python-dotenv>=1.0.1
click>=8.1.7
